Do not use a linker script when building the libretro core.
On Windows, a linker script is not needed because Windows requires a
library to use "dllexport" on functions in its public API. On macOS,
clang does not support a linker script (or at least, not like this),
so we can't use it. On Linux it works, but it's mostly only useful for
libretro hosts that link directly against the library, not ones that
load cores dynamically.

Since higan is too resource intensive to be much use on the kinds
of platforms that *do* link directly with libretro cores (jailbroken
consoles, mostly), let's just drop the linker script.

See https://board.byuu.org/viewtopic.php?p=47768#p47768 and following
for details.