Make the "reset" function do a soft-reset.
When the libretro interface was written, higan (temporarily) did not
have a soft-reset feature, so the reset function was wired up to the
"power-cycle" feature. A proper soft-reset was re-added in higan v106,
and most people that care would expect "reset" to be a soft-reset
(like the physical "Reset" button on a real Super Famicom), so let's do
that instead.